chore: Refactored code and split BillsListView
