+++ src/views/Bills/components/BillsListView/components/BillsListMenu.tsx
@@ -3,10 +3,10 @@ import React, { useState } from 'react'
 import { Checkbox, Flex, Input, Svg, Text, useMatchBreakpoints, NetworkButton, Toggle } from '@ape.swap/uikit'
 import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import { useTranslation } from 'contexts/Localization'
-import { SearchText, styles } from './styles'
+import { SearchText, styles } from '../styles'
 import useSelectNetwork from 'hooks/useSelectNetwork'
 import { AVAILABLE_CHAINS_ON_LIST_VIEW_PRODUCTS, LIST_VIEW_PRODUCTS } from 'config/constants/chains'
-import { BillsListMenuProps, FILTER_OPTIONS, SORT_OPTIONS } from './types'
+import { BillsListMenuProps, FILTER_OPTIONS, SORT_OPTIONS } from '../types'
 import MenuSelect from './MenuSelect'
 import { AnimatePresence, motion } from 'framer-motion'
