+++ src/views/Bills/components/UserBillsView/components/UserBillsMenu.tsx
@@ -3,11 +3,11 @@ import React, { useState } from 'react'
 import { Checkbox, Flex, Text, Input, useMatchBreakpoints, Toggle, Svg } from '@ape.swap/uikit'
 import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import { useTranslation } from 'contexts/Localization'
-import { ClaimAllWrapper, SearchText, styles } from './styles'
-import ClaimAll from '../Actions/ClaimAll'
+import { ClaimAllWrapper, SearchText, styles } from '../styles'
+import ClaimAll from '../../Actions/ClaimAll'
 import { Bills } from 'state/types'
-import { FILTER_OPTIONS, SORT_OPTIONS } from './types'
-import MenuSelect from '../BillsListView/MenuSelect'
+import { FILTER_OPTIONS, SORT_OPTIONS } from '../types'
+import MenuSelect from '../../BillsListView/components/MenuSelect'
 import { AnimatePresence, motion } from 'framer-motion'
 
 export interface UserBillsMenuProps {
